+++ Project/Assets/Scripts/Player.cs
@@ -17,6 +17,8 @@ public class Player : MonoBehaviour
 
     float xSensitivity = 3f, ySensitivity = 3f;
 
+    bool cursorLock = true;
+
     //向きの制限
     float minX = -90f, maxX = 90f;
 
@@ -57,6 +59,7 @@ public class Player : MonoBehaviour
         cam.transform.localRotation = cameraRot;
         transform.localRotation = playerRot;
 
+        UpdareCursorLock();
     }
 
     private void FixedUpdate()
@@ -92,6 +95,27 @@ public class Player : MonoBehaviour
         forward.y = 0.0f;
 
         transform.position += forward * z + cam.transform.right * x;
+    }
+
+    private void UpdareCursorLock()
+    {
+        if (Input.GetKeyDown(KeyCode.Escape))
+        {
+            cursorLock = false;
+        }
+        else if (Input.GetMouseButton(0))
+        {
+            cursorLock = true;
+        }
+
+        if (cursorLock)
+        {
+            Cursor.lockState = CursorLockMode.Locked;
+        }
+        else
+        {
+            Cursor.lockState = CursorLockMode.None;
+        }
     }
 
     public Quaternion ClampRotation(Quaternion quat)
