--- Project/Assets/Scripts/ScriptableObject.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c9a2cd14b4597ff438980e98c08a5c5d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
