--- Project/Assets/Scripts/ScriptableObject/Pistol.asset
@@ -1,21 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: d80864f4571f9444fa7946d68c2ff9f3, type: 3}
-  m_Name: Pistol
-  m_EditorClassIdentifier: 
-  weaponName: Pistol
-  level: 010000000200000003000000
-  damage: 28000000320000003c000000
-  headShot: 500000006400000078000000
-  cost: 000000006400000096000000
-  rate: 000000000000000000000000
-  add: 000000000000000000000000
