--- Project/Assets/Scripts/ScriptableObject/Pistol.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: eca4d92d9c990d84ab93b7b5bca71043
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
