--- Project/Assets/Scripts/ScriptableObject/WeaponData.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d80864f4571f9444fa7946d68c2ff9f3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
