--- Project/Assets/Scripts/UI/ButtonEx.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1f23a0b9da5a57444a6d5205e647ca16
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
