--- Project/Assets/Scripts/UI/PurchaseUI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3cb088f228182f4428ce12d58bb70615
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
