--- Project/Assets/Scripts/UI/UIManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d09207dd21e793b4f9b8bc35bcdadac0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
