--- Project/Assets/Scripts/UI/UpdateUI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4502b60d6f0e5d54b94703dd04a06ad7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
